FSPT-564: moves choosing a form type into the managed helper
Updates the support question types into a useful mapping to the type of
form to use to configure that managed expression - this should
necessarily keep this up to date and not be an extra configuration to
make sure is in line.

Moves forms and types around to make the order of imports and related
code sensible.
