+++ app/developers/forms.py
@@ -1,11 +1,9 @@
-from typing import TYPE_CHECKING, Any, Mapping, Sequence
+from typing import TYPE_CHECKING
 
 from flask_wtf import FlaskForm
-from govuk_frontend_wtf.wtforms_widgets import GovRadioInput, GovSelect, GovSubmitInput, GovTextInput
-from wtforms import IntegerField, RadioField, SelectField, SubmitField
+from govuk_frontend_wtf.wtforms_widgets import GovRadioInput, GovSelect, GovSubmitInput
+from wtforms import RadioField, SelectField, SubmitField
 from wtforms.validators import DataRequired, Optional
-
-from app.common.expressions.managed import ManagedExpressions
 
 if TYPE_CHECKING:
     from app.common.data.models import Question
@@ -53,25 +51,3 @@ class ConditionSelectQuestionForm(FlaskForm):
 
     def add_question_options(self, questions: list["Question"]) -> None:
         self.question.choices = [(question.id, f"{question.text} ({question.name})") for question in questions]
-
-
-class AddNumberConditionForm(FlaskForm):
-    # todo: should any condition or validation be able to override the human readable message I suspect so
-    type = RadioField(
-        "Only show the question if the answer is",
-        choices=[(ManagedExpressions.GREATER_THAN, ManagedExpressions.GREATER_THAN.value)],
-        validators=[DataRequired("Select what the answer should be to show this question")],
-        widget=GovRadioInput(),
-    )
-    value = IntegerField("Value", widget=GovTextInput(), validators=[Optional()])
-
-    submit = SubmitField("Add condition", widget=GovSubmitInput())
-
-    def validate(self, extra_validators: Mapping[str, Sequence[Any]] | None = None) -> bool:
-        # fixme: only validate the value if the type has been set, there's probably
-        #        a better way to do this
-        if self.type.data:
-            self.value.validators = [DataRequired("Enter a value")]
-
-        # fixme: IDE realises this is a FlaskForm and bool but mypy is calling it "Any" on pre-commit
-        return super().validate(extra_validators=extra_validators)  # type: ignore
